Added patch to newcalcal.csh to make it 'bug-compatible' with gpcopy. Also found out that I can't spell the word 'isn't'
git-svn-id: https://svn.hcro.org/mmm@333 9b194dde-2b4d-0410-98f0-bba1668eb47e

+++ karto/RAPIDBeta/newautomap.csh
@@ -1047,7 +1047,7 @@ echo " "
 goto invert
 
 enderr:
-rm -r $wd
+if ($debug) rm -r $wd
 exit 1
 
 failsafe:
